change type of cell size to float

+++ include/info.hpp
@@ -1,3 +1,3 @@
 const unsigned int Width = 20;
 const unsigned int Height = 20;
-const unsigned int CellSize = 30;
+const float CellSize = 30;
